+++ FlyShelf_PC/Classes/DevicePairingManager.cs
@@ -27,7 +27,7 @@ namespace FlyShelf.Classes
         public string localUrl { get; set; } = "";
         public string globalUrl { get; set; } = "";
         public string pin { get; set; } = "";
-        public long timestamp { get; set; }
+        public double timestamp { get; set; }
     }
 
     public class PairedDevice
@@ -479,7 +479,7 @@ namespace FlyShelf.Classes
                 // Check if code is still fresh (5 min TTL)
                 if (info != null && info.timestamp > 0)
                 {
-                    long ageMs = NetworkClock.UtcNowMs - info.timestamp;
+                    double ageMs = NetworkClock.UtcNowMs - info.timestamp;
                     if (ageMs > 5 * 60_000)
                     {
                         Logger.LogAction("PAIR CODE", $"Code {upperCode} expired ({ageMs / 1000}s old)");
